requirements.txt
httpx
fastmcp
langchain
langchain-openai
langchain-text-splitters
pymilvus
python-dotenv
pyyaml
